pyproject.toml
[tool.poetry]
name = "GasPy"
version = "0.1.0"
description = "A Django implementation for GAS"
authors = ["Fundor333 <matteo.scarpa@fundor333.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.8"
reportlab = "^3.5.34"
gunicorn = "^20.0.4"
psycopg2 = "^2.8.4"
django = "^3.0.4"
django-guardian = "^2.2.0"
django-filter = "^2.2.0"
openpyxl = "^3.0.3"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5"
black = "^19.10b0"
pylint-django = "^2.0.14"
pylint = "^2.4.4"
django-role-permissions = "^2.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
